Move otel tracing helper stuff to _shared
This stuff is not needed in airflow core.  It is not public interface.  And it may at some point be needed by task sdk.
