+++ airflow-core/src/airflow/__init__.py
@@ -92,7 +92,7 @@ __lazy_imports: dict[str, tuple[str, str, bool]] = {
     "Stats": (".observability.stats", "Stats", True),
     "Trace": (".observability.trace", "Trace", True),
     "metrics": (".observability.metrics", "", True),
-    "traces": (".observability.traces", "", True),
+    "traces": ("._shared.observability.traces", "", True),
 }
 if TYPE_CHECKING:
     # These objects are imported by PEP-562, however, static analyzers and IDE's
